+++ src/components/newsListingCard/NewsListingCard.jsx
@@ -11,7 +11,7 @@ const NewsListingCard = ({ date, heading, description, category }) => {
   const dateString = date.toLocaleString();
   const formatDate = (dateString) => {
     const date = new Date(dateString);
-    const options = { day: "2-digit", month: "short", year: "numeric" };
+    const options = { day: "2-digit", month: "long", year: "numeric" };
     return date.toLocaleDateString("en-GB", options);
   };
   const [expanded, setExpanded] = useState(false);
